+++ cg/cli/workflow/microsalt/base.py
@@ -58,11 +58,13 @@ microsalt.add_command(resolve_compression)
 @microsalt.command()
 @OPTION_TICKET
 @OPTION_SAMPLE
+@OPTION_DRY_RUN
 @ARGUMENT_UNIQUE_IDENTIFIER
 @click.pass_obj
-def link(context: CGConfig, ticket: bool, sample: bool, unique_id: str) -> None:
+def link(context: CGConfig, ticket: bool, sample: bool, unique_id: str, dry_run: bool) -> None:
     """Link microbial FASTQ files to dedicated analysis folder for a given case, ticket or sample"""
-
+    if dry_run:
+        return
     analysis_api: MicrosaltAnalysisAPI = context.meta_apis["analysis_api"]
     case_id, sample_id = analysis_api.resolve_case_sample_id(
         sample=sample, ticket=ticket, unique_id=unique_id
@@ -82,19 +84,25 @@ def link(context: CGConfig, ticket: bool, sample: bool, unique_id: str) -> None:
 def config_case(
     context: CGConfig, dry_run: bool, ticket: bool, sample: bool, unique_id: str
 ) -> None:
-    """ Create a config file for a case or a sample analysis in microSALT """
+    """Create a config file for a case or a sample analysis in microSALT"""
 
     analysis_api: MicrosaltAnalysisAPI = context.meta_apis["analysis_api"]
     case_id, sample_id = analysis_api.resolve_case_sample_id(
         sample=sample, ticket=ticket, unique_id=unique_id
     )
-    sample_objs = analysis_api.get_samples(case_id=case_id, sample_id=sample_id)
+    sample_objs: List[models.Sample] = analysis_api.get_samples(
+        case_id=case_id, sample_id=sample_id
+    )
 
     if not sample_objs:
         LOG.error("No sample found for that ticket/sample_id")
         raise click.Abort
 
-    parameters: List[dict] = [analysis_api.get_parameters(sample_obj) for sample_obj in sample_objs]
+    parameters: List[dict] = [
+        analysis_api.get_parameters(sample_obj)
+        for sample_obj in sample_objs
+        if sample_obj.sequencing_qc
+    ]
     filename: str = sample_id or case_id
     config_case_path: Path = analysis_api.get_config_path(filename=filename)
     if dry_run:
@@ -127,7 +135,7 @@ def run(
     sample: bool,
     unique_id: Any,
 ) -> None:
-    """ Start microSALT workflow by providing case, ticket or sample id """
+    """Start microSALT workflow by providing case, ticket or sample id"""
 
     analysis_api: MicrosaltAnalysisAPI = context.meta_apis["analysis_api"]
     case_id, sample_id = analysis_api.resolve_case_sample_id(
@@ -152,12 +160,10 @@ def run(
             "--input",
             Path(fastq_path, sample_id).absolute().as_posix(),
         ]
-    analysis_api.process.run_command(parameters=analyse_command, dry_run=dry_run)
 
     if sample_id or dry_run:
+        analysis_api.process.run_command(parameters=analyse_command, dry_run=dry_run)
         return
-
-    analysis_api.set_statusdb_action(case_id=case_id, action="running")
     try:
         analysis_api.add_pending_trailblazer_analysis(case_id=case_id)
     except Exception as e:
@@ -166,6 +172,13 @@ def run(
             case_id,
             e.__class__.__name__,
         )
+    try:
+        analysis_api.set_statusdb_action(case_id=case_id, action="running")
+        analysis_api.process.run_command(parameters=analyse_command, dry_run=dry_run)
+    except:
+        LOG.error("Failed to run analysis!")
+        analysis_api.set_statusdb_action(case_id=case_id, action=None)
+        raise
 
 
 @microsalt.command()
@@ -179,9 +192,8 @@ def start(
 ) -> None:
     """Start whole microSALT workflow by providing case, ticket or sample id"""
     LOG.info("Starting Microsalt workflow for %s", unique_id)
-    if not sample and not ticket:
-        context.invoke(resolve_compression, case_id=unique_id, dry_run=dry_run)
-    context.invoke(link, ticket=ticket, sample=sample, unique_id=unique_id)
+
+    context.invoke(link, ticket=ticket, sample=sample, unique_id=unique_id, dry_run=dry_run)
     context.invoke(config_case, ticket=ticket, sample=sample, unique_id=unique_id, dry_run=dry_run)
     context.invoke(run, ticket=ticket, sample=sample, unique_id=unique_id, dry_run=dry_run)