+++ tests/cli/workflow/microsalt/conftest.py
@@ -11,7 +11,7 @@ from tests.store_helpers import StoreHelpers
 
 
 class MockLims:
-    """ provides a mock class overriding relevant methods for microbial cli """
+    """provides a mock class overriding relevant methods for microbial cli"""
 
     def __init__(self):
         self.lims = self
@@ -19,17 +19,17 @@ class MockLims:
         self.lims_sample = None
 
     def sample(self, sample_id: str):
-        """ return a mock sample """
+        """return a mock sample"""
 
         class LimsSample:
-            """ A mock class for a sample coming from LIMS. It only needs a comment """
+            """A mock class for a sample coming from LIMS. It only needs a comment"""
 
             def __init__(self, sample_id):
                 self.sample_id = sample_id
                 self.sample_data = {"comment": "a comment in LimsSample"}
 
             def get(self, key):
-                """ only here to get the sample.get('comment') """
+                """only here to get the sample.get('comment')"""
                 return self.sample_data.get(key, "")
 
         # haha, it's a factory!
@@ -39,12 +39,12 @@ class MockLims:
         return self.lims_sample
 
     def get_prep_method(self, sample_id: str) -> str:
-        """ Return a prep method name. Needs to be in format 'dddd:dd string' """
+        """Return a prep method name. Needs to be in format 'dddd:dd string'"""
         self.sample_id = sample_id
         return "1337:00 Test prep method"
 
     def get_sequencing_method(self, sample_id: str) -> str:
-        """ Return a sequencing method name. Needs to be in format 'dddd:dd string' """
+        """Return a sequencing method name. Needs to be in format 'dddd:dd string'"""
         self.sample_id = sample_id
         return "1338:00 Test sequencing method"
 
@@ -56,13 +56,13 @@ class MockLims:
 
 
 class LimsFactory:
-    """ Just give one LIMS """
+    """Just give one LIMS"""
 
     single_lims = None
 
     @classmethod
     def produce(cls):
-        """ Produce me one LIMS """
+        """Produce me one LIMS"""
         if not cls.single_lims:
             cls.single_lims = MockLims()
 
@@ -77,7 +77,7 @@ def lims_api():
 
 @pytest.fixture(scope="function")
 def base_context(cg_context: CGConfig, helpers: StoreHelpers, lims_api):
-    """ The click context for the microsalt cli """
+    """The click context for the microsalt cli"""
     helpers.add_microbial_sample(cg_context.status_db)
     cg_context.lims_api_ = lims_api
     cg_context.meta_apis["analysis_api"] = MicrosaltAnalysisAPI(cg_context)
@@ -86,19 +86,19 @@ def base_context(cg_context: CGConfig, helpers: StoreHelpers, lims_api):
 
 @pytest.fixture(name="microbial_sample_id")
 def fixture_microbial_sample_id():
-    """ Define a name for a microbial sample """
+    """Define a name for a microbial sample"""
     return "microbial_sample_id"
 
 
 @pytest.fixture(name="microbial_sample_name")
 def fixture_microbial_sample_name():
-    """ Define a name for a microbial sample """
+    """Define a name for a microbial sample"""
     return "microbial_sample_name"
 
 
 @pytest.fixture(name="microbial_ticket")
 def fixture_microbial_ticket():
-    """ Define a ticket for a microbial order """
+    """Define a ticket for a microbial order"""
     return "123456"