+++ tests/store/conftest.py
@@ -11,14 +11,14 @@ from cg.store import Store, models
 
 @pytest.fixture
 def application_versions_file(fixtures_dir: Path) -> str:
-    """"application version import file"""
+    """ "application version import file"""
     _file = fixtures_dir / "store/api/application_versions.xlsx"
     return str(_file)
 
 
 @pytest.fixture
 def applications_file(fixtures_dir):
-    """"application import file"""
+    """ "application import file"""
     _file = fixtures_dir / "store/api/applications.xlsx"
     return str(_file)