+++ CHANGELOG.md
@@ -10,13 +10,78 @@ Try to use the following format:
 
 ## [x.x.x]
 ### Added
-- gisaid uppload support via cli
-- gisaid API
-
 ### Changed
 ### Fixed
 
-## [20.22.0]
+## [21.4.2]
+
+### Fixed
+- Remove default option for mip-dna priority
+
+## [21.4.1]
+### Changed
+- Change how samples are fetched for cgstats select command
+### Fixed
+- Bug when fetching info with `cg demultiplex select`-command
+
+## [21.4.0]
+### Added
+- A column in customer-group admin view in to show customers in the group
+
+## [21.3.1]
+### Fixed
+- PDC backup destination server for 2500 flowcells
+
+## [21.3.0]
+### Changed
+- Remove dependency `cgstats` from requirements.txt and move used functionality into CG
+
+## [21.2.0]
+### Added
+- Functionality for the cgstats api in CG
+
+## [21.1.0]
+### Added
+Select analyses to be uploaded to vogue based on analysis completed date (before or after a date, of between two dates)
+Add uploaded to vogue date to analysis table
+Only select potential analyses to upload that have not been uploaded
+
+## [21.0.0]
+
+### Changed
+- Add support for balsamic 7.x.x
+- Rework Balsamic server configurations
+
+### Fixed
+- Upload to scout now possible for all analysis types through cg upload scout
+
+
+## [20.26.2]
+### Added
+- Added more DNA and RNA positive control cases to VALIDATION_CASES
+
+## [20.26.1]
+### Fixed
+- Workflow linking also links undetermined files when
+
+## [20.26.0]
+### Changed
+- Cases to analyze query checks if any samples in case are newer than latest analysis to start topups
+- Microsalt config-case wont include samples that dont pass sequencing qc
+
+## [20.25.0]
+### Changed
+- Changes that customer contact persons are referred to as emails instead of users. This removes the need to add each contact as a user in the system and it becomes easier to manage the user list
+
+## [20.24.0]
+### Added
+- Show customers in user admin view
+
+## [20.23.0]
+### Added
+- gisaid uppload support via cli
+- gisaid API
+
 ### Added
 - Add command `cg get analysis` to view analysis information
 
@@ -52,15 +117,12 @@ Try to use the following format:
 ### Fixed
 - Update sequenced at timestamp of sample whenever sample has been sequenced and flowcell transferred
 
-
 ## [20.19.2]
 
 ### Fixed
 - Bug when instantiating analysis api in upload
 
-
 ## [20.19.1]
-
 ### Fixed
 - Bug when fetching the api in mip workflow cli base